fix(product): reorganize gql queries for shopify
