+++ libs/product/driver/shopify/src/queries/gql/get-product.ts
@@ -1,7 +1,7 @@
 import { gql } from 'apollo-angular';
 
-import { ShopifyProductSingleResponse } from './response.types';
-import { ShopifyProductIDVariables } from './variables.types';
+import { ShopifyProductSingleResponse } from '../response.types';
+import { ShopifyProductIDVariables } from '../variables.types';
 
 export const DAFF_SHOPIFY_GET_PRODUCT_QUERY_NAME = 'ShopifyGetAProduct';
