+++ libs/product/driver/shopify/src/queries/gql/public_api.ts
@@ -1,0 +1,3 @@
+export * from './get-all-products';
+export * from './get-product-by-url';
+export * from './get-product';
